kurral-cli-main/kurral-cli-main/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kurral"
version = "0.1.1"
description = "Deterministic LLM agent testing and replay framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Kurral Team", email = "team@kurral.dev"}
]
keywords = ["llm", "testing", "determinism", "langsmith", "replay", "ai", "agents", "regression-testing", "a-b-testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.1.7",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.34.0",
    "httpx>=0.26.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tenacity>=8.2.3",
    "python-dateutil>=2.8.2",
]

[project.urls]
Homepage = "https://github.com/kurral-dev/kurral-cli"
Documentation = "https://github.com/kurral-dev/kurral-cli#readme"
Repository = "https://github.com/kurral-dev/kurral-cli"
Issues = "https://github.com/kurral-dev/kurral-cli/issues"
Changelog = "https://github.com/kurral-dev/kurral-cli/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
kurral = "kurral.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["kurral*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=kurral --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

